+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fix `BakeChapterTitle` usage in `algebra-1`
 * Change Algebra 1 locale of `Link to Learning` to `Classroom Connection`
 * Ensure numbered exercises are only baked once in `algebra-1`
+* Bake `document` notes in `algebra-1`
 
 ## [v2.26.1] - 2025-05-22